+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Method `set_params` to change parameters of ETNA objects [#1102](https://github.com/tinkoff-ai/etna/pull/1102)
+- Function `plot_forecast_decomposition` [#1129](https://github.com/tinkoff-ai/etna/pull/1129)
 -
 ### Changed